+++ src/migration/ember-addon/index.js
@@ -1,5 +1,5 @@
 import { updatePaths } from '../../utils/ember-addon/app/components.js';
-import { moveFiles } from '../../utils/move-files.js';
+import { moveFiles } from '../../utils/files.js';
 import { migrationStrategyForAddonFolder } from './addon/index.js';
 import { migrationStrategyForAppFolder } from './app/index.js';
 import { migrationStrategyForTestsFolder } from './tests/index.js';
