+++ src/migration/ember-app/app/component-classes.js
@@ -2,7 +2,7 @@ import { join } from 'node:path';
 
 import glob from 'glob';
 
-import { mapPaths } from '../../../utils/map-paths.js';
+import { mapFilePath } from '../../../utils/files.js';
 
 export function migrationStrategyForComponentClasses(options) {
   const { podPath, projectRoot } = options;
@@ -16,7 +16,7 @@ export function migrationStrategyForComponentClasses(options) {
 
   return oldPaths.map((oldPath) => {
     if (oldPath.endsWith('.d.ts')) {
-      return mapPaths(oldPath, {
+      return mapFilePath(oldPath, {
         find: {
           directory: join('app', podPath, 'components'),
           file: 'component.d.ts',
@@ -28,7 +28,7 @@ export function migrationStrategyForComponentClasses(options) {
     }
 
     if (oldPath.endsWith('.ts')) {
-      return mapPaths(oldPath, {
+      return mapFilePath(oldPath, {
         find: {
           directory: join('app', podPath, 'components'),
           file: 'component.ts',
@@ -39,7 +39,7 @@ export function migrationStrategyForComponentClasses(options) {
       });
     }
 
-    return mapPaths(oldPath, {
+    return mapFilePath(oldPath, {
       find: {
         directory: join('app', podPath, 'components'),
         file: 'component.js',
