+++ src/migration/ember-app/app/route-templates.js
@@ -2,7 +2,7 @@ import { join } from 'node:path';
 
 import glob from 'glob';
 
-import { mapPaths } from '../../../utils/map-paths.js';
+import { mapFilePath } from '../../../utils/files.js';
 
 export function migrationStrategyForRouteTemplates(options) {
   const { podPath, projectRoot } = options;
@@ -15,7 +15,7 @@ export function migrationStrategyForRouteTemplates(options) {
   );
 
   return oldPaths.map((oldPath) => {
-    return mapPaths(oldPath, {
+    return mapFilePath(oldPath, {
       find: {
         directory: join('app', podPath),
         file: 'template.hbs',
