+++ src/migration/ember-app/tests/route-controllers.js
@@ -2,7 +2,7 @@ import { join } from 'node:path';
 
 import glob from 'glob';
 
-import { mapPaths } from '../../../utils/map-paths.js';
+import { mapFilePath } from '../../../utils/files.js';
 
 export function migrationStrategyForRouteControllers(options) {
   const { podPath, projectRoot } = options;
@@ -25,7 +25,7 @@ export function migrationStrategyForRouteControllers(options) {
 
   const newPaths1 = oldPaths1.map((oldPath) => {
     if (oldPath.endsWith('.ts')) {
-      return mapPaths(oldPath, {
+      return mapFilePath(oldPath, {
         find: {
           directory: join('tests/unit', podPath),
           file: 'controller-test.ts',
@@ -36,7 +36,7 @@ export function migrationStrategyForRouteControllers(options) {
       });
     }
 
-    return mapPaths(oldPath, {
+    return mapFilePath(oldPath, {
       find: {
         directory: join('tests/unit', podPath),
         file: 'controller-test.js',
@@ -59,7 +59,7 @@ export function migrationStrategyForRouteControllers(options) {
 
   const newPaths2 = oldPaths2.map((oldPath) => {
     if (oldPath.endsWith('.ts')) {
-      return mapPaths(oldPath, {
+      return mapFilePath(oldPath, {
         find: {
           directory: join('tests/unit', podPath, 'controllers'),
           file: 'controller-test.ts',
@@ -70,7 +70,7 @@ export function migrationStrategyForRouteControllers(options) {
       });
     }
 
-    return mapPaths(oldPath, {
+    return mapFilePath(oldPath, {
       find: {
         directory: join('tests/unit', podPath, 'controllers'),
         file: 'controller-test.js',
