+++ src/migration/ember-engine/addon/component-classes.js
@@ -1,6 +1,6 @@
 import glob from 'glob';
 
-import { mapPaths } from '../../../utils/map-paths.js';
+import { mapFilePath } from '../../../utils/files.js';
 
 export function migrationStrategyForComponentClasses(options) {
   const { projectRoot } = options;
@@ -11,7 +11,7 @@ export function migrationStrategyForComponentClasses(options) {
 
   return oldPaths.map((oldPath) => {
     if (oldPath.endsWith('.d.ts')) {
-      return mapPaths(oldPath, {
+      return mapFilePath(oldPath, {
         find: {
           directory: 'addon/components',
           file: 'component.d.ts',
@@ -23,7 +23,7 @@ export function migrationStrategyForComponentClasses(options) {
     }
 
     if (oldPath.endsWith('.ts')) {
-      return mapPaths(oldPath, {
+      return mapFilePath(oldPath, {
         find: {
           directory: 'addon/components',
           file: 'component.ts',
@@ -34,7 +34,7 @@ export function migrationStrategyForComponentClasses(options) {
       });
     }
 
-    return mapPaths(oldPath, {
+    return mapFilePath(oldPath, {
       find: {
         directory: 'addon/components',
         file: 'component.js',
