+++ src/migration/ember-engine/tests/route-routes.js
@@ -1,6 +1,6 @@
 import glob from 'glob';
 
-import { mapPaths } from '../../../utils/map-paths.js';
+import { mapFilePath } from '../../../utils/files.js';
 
 export function migrationStrategyForRouteRoutes(options) {
   const { projectRoot } = options;
@@ -14,7 +14,7 @@ export function migrationStrategyForRouteRoutes(options) {
 
   const newPaths1 = oldPaths1.map((oldPath) => {
     if (oldPath.endsWith('.ts')) {
-      return mapPaths(oldPath, {
+      return mapFilePath(oldPath, {
         find: {
           directory: 'tests/unit',
           file: 'route-test.ts',
@@ -25,7 +25,7 @@ export function migrationStrategyForRouteRoutes(options) {
       });
     }
 
-    return mapPaths(oldPath, {
+    return mapFilePath(oldPath, {
       find: {
         directory: 'tests/unit',
         file: 'route-test.js',
@@ -45,7 +45,7 @@ export function migrationStrategyForRouteRoutes(options) {
 
   const newPaths2 = oldPaths2.map((oldPath) => {
     if (oldPath.endsWith('.ts')) {
-      return mapPaths(oldPath, {
+      return mapFilePath(oldPath, {
         find: {
           directory: 'tests/unit/routes',
           file: 'route-test.ts',
@@ -56,7 +56,7 @@ export function migrationStrategyForRouteRoutes(options) {
       });
     }
 
-    return mapPaths(oldPath, {
+    return mapFilePath(oldPath, {
       find: {
         directory: 'tests/unit/routes',
         file: 'route-test.js',
