--- src/utils/map-paths.js
@@ -1,15 +1,0 @@
-export function mapPaths(oldPath, { find, replace }) {
-  const { directory, file } = find;
-
-  if (!oldPath.endsWith(file)) {
-    return;
-  }
-
-  const key = oldPath
-    .replace(new RegExp(`^${directory}/`), '')
-    .replace(new RegExp(`/${file}$`), '');
-
-  const newPath = replace(key);
-
-  return [oldPath, newPath];
-}
